test meta round trip for file table
